+++ EcoMarket/src/main/java/EcoMarket/EcoMarket/Productos/repository/ProveedorRepository.java
@@ -8,6 +8,6 @@ import EcoMarket.EcoMarket.Productos.model.Proveedor;
 
 
 @Repository
-public interface ProveedorRepository extends JpaRepository<Proveedor, Long> {
+public interface ProveedorRepository extends JpaRepository<Proveedor, Integer> {
 
 }
