+++ EcoMarket/src/main/java/EcoMarket/EcoMarket/Productos/service/ProductoService.java
@@ -5,8 +5,10 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import Productos.model.Producto;
-import Productos.repository.ProductoRepository;
+import EcoMarket.EcoMarket.Productos.model.Producto;
+import EcoMarket.EcoMarket.Productos.repository.ProductoRepository;
+
+
 
 @Service
 public class ProductoService {
@@ -18,7 +20,7 @@ public class ProductoService {
         return productoRepository.findAll();
     }
 
-    public Producto obtenerPorId(Long id) {
+    public Producto obtenerPorId(int id) {
         return productoRepository.findById(id).orElse(null);
     }
 
@@ -26,7 +28,7 @@ public class ProductoService {
         return productoRepository.save(producto);
     }
 
-    public void eliminar(Long id) {
+    public void eliminar(int id) {
         productoRepository.deleteById(id);
     }
 }
