+++ EcoMarket/src/main/java/EcoMarket/EcoMarket/Usuarios/repository/PermisoRepository.java
@@ -8,7 +8,7 @@ import EcoMarket.EcoMarket.Usuarios.model.Permiso;
 
 
 @Repository
-public interface  PermisoRepository extends JpaRepository<Permiso, Long> {
+public interface  PermisoRepository extends JpaRepository<Permiso, Integer> {
 
 
 }
