+++ EcoMarket/src/main/java/EcoMarket/EcoMarket/Usuarios/service/PermisoService.java
@@ -5,8 +5,10 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import Usuarios.model.Permiso;
-import Usuarios.repository.PermisoRepository;
+import EcoMarket.EcoMarket.Usuarios.model.Permiso;
+import EcoMarket.EcoMarket.Usuarios.repository.PermisoRepository;
+
+
 
 
 @Service
@@ -19,7 +21,7 @@ public class PermisoService {
         return permisoRepository.findAll();
     }
 
-    public Permiso ObtenerPorId(Long id) {
+    public Permiso ObtenerPorId(int id) {
         return permisoRepository.findById(id).orElse(null);
     }
 
@@ -27,7 +29,7 @@ public class PermisoService {
         return permisoRepository.save(permiso);
     }
 
-    public void eliminar(Long id) {
+    public void eliminar(int id) {
         permisoRepository.deleteById(id);
     }
