+++ EcoMarket/src/main/java/EcoMarket/EcoMarket/Usuarios/service/RolService.java
@@ -5,8 +5,10 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import Usuarios.model.Rol;
-import Usuarios.repository.RolRepository;
+import EcoMarket.EcoMarket.Usuarios.model.Rol;
+import EcoMarket.EcoMarket.Usuarios.repository.RolRepository;
+
+
 
 @Service
 public class RolService {
@@ -18,7 +20,7 @@ public class RolService {
         return rolRepository.findAll();
     }
 
-    public Rol ObtenerPorId(Long id) {
+    public Rol ObtenerPorId(int id) {
         return rolRepository.findById(id).orElse(null);
     }
 
@@ -26,7 +28,7 @@ public class RolService {
         return rolRepository.save(rol);
     }
 
-    public void eliminar(Long id) {
+    public void eliminar(int id) {
         rolRepository.deleteById(id);
     }
 }
